Split pkg/util/crypt key and device functionality
`pkg/util/crypt` contained code that calls out to `cryptsetup`, which is
found via routines that rely on a valid buildcfg (only present when
SingularityCE is built via `mconfig`). This, in turn, prevents
`pkg/build/types` being imported from other projects as `bundle.go`
imports `pkg/util/crypt`.

Move the device handling portions of the crypt code to
`internal/pkg/util/crypt`.

Keep the remaining key handling portions of the code, in
`pkg/util/cryptkey`. These now have no internal or buildcfg
dependencies.

Fixes sylabs/singularity#181
