+++ cmd/internal/cli/actions_linux.go
@@ -35,7 +35,7 @@ import (
 	singularityConfig "github.com/hpcng/singularity/pkg/runtime/engine/singularity/config"
 	"github.com/hpcng/singularity/pkg/sylog"
 	"github.com/hpcng/singularity/pkg/util/capabilities"
-	"github.com/hpcng/singularity/pkg/util/crypt"
+	"github.com/hpcng/singularity/pkg/util/cryptkey"
 	"github.com/hpcng/singularity/pkg/util/fs/proc"
 	"github.com/hpcng/singularity/pkg/util/gpu"
 	"github.com/hpcng/singularity/pkg/util/namespaces"
@@ -396,7 +396,7 @@ func execStarter(cobraCmd *cobra.Command, image string, args []string, name stri
 				sylog.Fatalf("Cannot load key for decryption: %v", err)
 			}
 
-			plaintextKey, err := crypt.PlaintextKey(keyInfo, engineConfig.GetImage())
+			plaintextKey, err := cryptkey.PlaintextKey(keyInfo, engineConfig.GetImage())
 			if err != nil {
 				sylog.Errorf("Cannot decrypt %s: %v", engineConfig.GetImage(), err)
 				sylog.Fatalf("Please check you are providing the correct key for decryption")
