+++ internal/pkg/runtime/engine/singularity/cleanup_linux.go
@@ -16,12 +16,12 @@ import (
 
 	"github.com/hpcng/singularity/internal/pkg/instance"
 	fakerootConfig "github.com/hpcng/singularity/internal/pkg/runtime/engine/fakeroot/config"
+	"github.com/hpcng/singularity/internal/pkg/util/crypt"
 	"github.com/hpcng/singularity/internal/pkg/util/priv"
 	"github.com/hpcng/singularity/internal/pkg/util/starter"
 	"github.com/hpcng/singularity/pkg/runtime/engine/config"
 	"github.com/hpcng/singularity/pkg/sylog"
 	"github.com/hpcng/singularity/pkg/util/capabilities"
-	"github.com/hpcng/singularity/pkg/util/crypt"
 )
 
 // CleanupContainer is called from master after the MonitorContainer returns.
